+++ lms/djangoapps/commerce/api/v0/urls.py
@@ -1,0 +1,16 @@
+""" API v0 URLs. """
+from django.conf.urls import patterns, url, include
+
+from commerce.api.v0 import views
+
+
+BASKET_URLS = patterns(
+    '',
+    url(r'^$', views.BasketsView.as_view(), name='create'),
+    url(r'^{}/order/$'.format(r'(?P<basket_id>[\w]+)'), views.BasketOrderView.as_view(), name='retrieve_order'),
+)
+
+urlpatterns = patterns(
+    '',
+    url(r'^baskets/', include(BASKET_URLS, namespace='baskets')),
+)
